hundle leaks !

+++ bonus_srcs/child.c
@@ -6,7 +6,7 @@
 /*   By: baouragh <baouragh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/07 16:14:27 by baouragh          #+#    #+#             */
-/*   Updated: 2024/02/08 21:31:50 by baouragh         ###   ########.fr       */
+/*   Updated: 2024/02/09 16:22:23 by baouragh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,13 +19,21 @@ void	child(t_fd fd, char *argv, char **env, int mod)
 
 	open_pipe(pfd);
 	id = fork();
+	if (id < 0 )
+	{
+		print_err("pipex: error occuerd with fork!",NULL);
+		exit(EXIT_FAILURE);
+	}
 	if (id == 0)
 	{
+		// fprintf(stderr,"child pid: %d\n",getpid());
 		fd_duper(fd, mod, pfd);
+		// while(1);
 		call_execev(env, argv);
 	}
 	else
 	{
+		wait(NULL);
 		close(pfd[1]);
 		dup_2(pfd[0], 0, 3);
 		unlink("tmp.txt");
